create a gatewayProtocol folder
